requirements.txt
flask
BeautifulSoup4
ntplib
feedparser
python-dateutil
pymysql
requests
